Add `copy`, `externalCopy`, and `release` options
**This is a semver major change**. This affects the behavior of
`Reference.get`, `Reference.apply`, and their sync/ignored counterparts.
These functions will now transfer the value out instead of
unconditionally wrapping it in a `Reference` instance (in the case of
`get`) or throwing (in the cast of `apply`). This has no effect if the
value is not transferable (e.g. calling `get` to get a reference to a
function).