Add Algolia to the list of users
We've been using `isolated-vm` at Algolia for our crawler (which I've been a dev on for the entire project life) for the past year and I have to say that it was a blast. It's one of our core security mechanisms.
We spent so much time looking at alternatives, that when we discovered it, it almost felt too good to be true.

By the way, `How to run untrusted JavaScript code` is a banned question on StackOverflow because it has been asked so many times. I've mentioned `isolated-vm` in one of the most popular ones and flagged a lot of them as duplicates.

I plan to do a few presentations about our usage of the lib, and would be happy to see our name in your user list. :)