Windows support
I thought I was close to getting this to build on MSVC but at the last
minute it choked hard on the value template stuff, because of course it
would. It's not a big deal to compile it with LLVM for Windows though.

The compiled binary is only good on the version of node it was compiled
against. That is a problem for another day.