More robust agent resource management
The ownership of an agent is a bit more clear now. The handle holds a
`weak_ptr` along with a shared `shared_ptr` which can be severed by the
agent host.

This causes most of the tests to fail because `v8::Global`s are being
reset after the agent is disposed. That's to be expected without changes
to remote handles.