Prevent move construction of tuple `visit`
You can't construct tuples in place but with a sneaky cast operator you
can pull it off in most cases.

It's not a matter of efficiency, there is a correctness concern.
Stateful visitors (key acceptors) should not be moved.

I guess one advantage of packed base classes is that they are easier to
construct. Of course in many of these visitors you can't get around it
because the types may not be unique.