Fix builds
Most broken builds are fixed. Windows is disabled, since it seems like
cross-build will be better there.

The absl mystery from two commits ago was because of symbols from nodejs
interfering with z8.

It crashes on basically everything that's not linux arm64, which is what
I've been testing on this whole time.